abstract NotificationConfigManager to support both MongoDB and Zookeeper
Added an abstraction class to allow using the NotificationConfigManager
with both MongoDB and Zookeeper.

Issue: BB-497
