+++ extensions/notification/NotificationConfigValidator.js
@@ -28,6 +28,7 @@ const joiSchema = joi.object({
     topic: joi.string(),
     monitorNotificationFailures: joi.boolean().default(true),
     notificationFailedTopic: joi.string().optional(),
+    zookeeperPath: joi.string().optional(),
     queueProcessor: {
         groupId: joi.string().required(),
         concurrency: joi.number().greater(0).default(1000),
