+++ extensions/notification/queueProcessor/QueueProcessor.js
@@ -4,7 +4,6 @@ const { EventEmitter } = require('events');
 const Logger = require('werelogs').Logger;
 const async = require('async');
 const assert = require('assert');
-const util = require('util');
 const { ZenkoMetrics } = require('arsenal').metrics;
 const errors = require('arsenal').errors;
 
@@ -34,6 +33,7 @@ class QueueProcessor extends EventEmitter {
      *
      * @constructor
      * @param {Object} mongoConfig - mongodb connnection configuration object
+     * @param {Object} zkConfig - zookeeper configuration object
      * @param {Object} kafkaConfig - kafka configuration object
      * @param {string} kafkaConfig.hosts - list of kafka brokers
      *   as "host:port[,host:port...]"
@@ -65,10 +65,11 @@ class QueueProcessor extends EventEmitter {
      * @param {String} destinationId - resource name/id of destination
      * @param {Object} destinationAuth - destination authentication config
      */
-    constructor(mongoConfig, kafkaConfig, notifConfig, destinationId,
+    constructor(mongoConfig, zkConfig, kafkaConfig, notifConfig, destinationId,
         destinationAuth) {
         super();
         this.mongoConfig = mongoConfig;
+        this.zkConfig = zkConfig;
         this.kafkaConfig = kafkaConfig;
         this.notifConfig = notifConfig;
         this.destinationId = destinationId;
@@ -84,10 +85,6 @@ class QueueProcessor extends EventEmitter {
         this.bnConfigManager = null;
         this._consumer = null;
         this._destination = null;
-        // Once the notification manager is initialized
-        // this will hold the callback version of the getConfig
-        // function of the notification config manager
-        this._getConfig = null;
 
         this.logger = new Logger('Backbeat:Notification:QueueProcessor');
     }
@@ -101,6 +98,10 @@ class QueueProcessor extends EventEmitter {
         try {
             this.bnConfigManager = new NotificationConfigManager({
                 mongoConfig: this.mongoConfig,
+                bucketMetastore: this.notifConfig.bucketMetastore,
+                zkClient: this.zkClient,
+                zkConfig: this.zkConfig,
+                zkPath: this.notifConfig.zookeeperPath,
                 logger: this.logger,
             });
             return this.bnConfigManager.setup(done);
@@ -181,9 +182,6 @@ class QueueProcessor extends EventEmitter {
                     this.emit('ready');
                     return next();
                 });
-                // callbackify getConfig from notification config manager
-                this._getConfig = util.callbackify(this.bnConfigManager
-                    .getConfig.bind(this.bnConfigManager));
                 return undefined;
             },
         ], err => {
@@ -231,7 +229,7 @@ class QueueProcessor extends EventEmitter {
         }
         const { bucket, key, eventType } = sourceEntry;
         try {
-            return this._getConfig(bucket, (err, notifConfig) => {
+            return this.bnConfigManager.getConfig(bucket, (err, notifConfig) => {
                 if (err) {
                     this.logger.error('Error while getting notification configuration', {
                         bucket,
