+++ lib/queuePopulator/QueuePopulator.js
@@ -199,8 +199,8 @@ class QueuePopulator {
                 return next(err);
             }),
             next => this._setupFailedCRRClients(next),
-            next => this._setupNotificationConfigManager(next),
-            next => this._setupZookeeper(err => {
+            next => this._setupZookeeper(next),
+            next => this._setupNotificationConfigManager(err => {
                 if (err) {
                     return next(err);
                 }
@@ -417,6 +417,8 @@ class QueuePopulator {
             try {
                 this.bnConfigManager = new NotificationConfigManager({
                     mongoConfig: this.qpConfig.mongo,
+                    bucketMetastore: this.extConfigs.notification.bucketMetastore,
+                    zkClient: this.zkClient,
                     logger: this.log,
                 });
                 return this.bnConfigManager.setup(done);
